requirements.txt
beautifulsoup4==4.14.3
langchain_community==0.4.1
langchain_core==1.2.6
langchain_huggingface==1.2.0
langchain_openai==1.1.6
matplotlib==3.10.8
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.52.2
tqdm==4.67.1
sentence-transformers==5.2.0
faiss-cpu==1.13.2